Fix script start.sh e dockerfile

+++ start.sh
@@ -1,12 +1,16 @@
 #!/bin/sh
-cd gateway
+cd tmp/gateway
 
+
+if [ ! -f /usr/src/gateway/gateway.jar -eq 1 ] || [ ! -f /usr/src/gateway/devicesimulator.jar -eq 2 ]; then 
 mvn clean package
 
 mv ./target/gateway*.jar ../../gateway.jar
 mv ./target/device*.jar ../../devicesimulator.jar
+
 cd ../..
 rm -rf tmp 
+fi
 
 java -jar /usr/src/gateway/gateway.jar &
 java -jar /usr/src/gateway/devicesimulator.jar
